+++ src/main/kotlin/at/asitplus/gradle/CommonDependencies.kt
@@ -22,7 +22,7 @@ internal inline fun KotlinDependencyHandler.addKotest(target: String? = null) {
 }
 
 @JvmOverloads
-inline fun kotest(module: String, target: String? = null) =
+ fun kotest(module: String, target: String? = null) =
     "io.kotest:kotest-$module${target.toSuffix()}:${AspVersions.kotest}"
 
 internal inline fun KotlinDependencyHandler.addKotestJvmRunner() {
@@ -31,25 +31,25 @@ internal inline fun KotlinDependencyHandler.addKotestJvmRunner() {
 }
 
 @JvmOverloads
-inline fun serialization(format: String, target: String? = null) =
+ fun serialization(format: String, target: String? = null) =
     "org.jetbrains.kotlinx:kotlinx-serialization-$format${target.toSuffix()}:${AspVersions.serialization}"
 
 @JvmOverloads
-inline fun ktor(module: String, target: String? = null) =
+ fun ktor(module: String, target: String? = null) =
     "io.ktor:ktor-$module${target.toSuffix()}:${AspVersions.ktor}"
 
 @JvmOverloads
-inline fun coroutines(target: String? = null) =
+ fun coroutines(target: String? = null) =
     "org.jetbrains.kotlinx:kotlinx-coroutines-core${target.toSuffix()}:${AspVersions.coroutines}"
 
 @JvmOverloads
-inline fun napier(target: String? = null) =
+ fun napier(target: String? = null) =
     "io.github.aakira:napier${target.toSuffix()}:${AspVersions.napier}"
 
 @JvmOverloads
-inline fun datetime(target: String? = null) =
+ fun datetime(target: String? = null) =
     "org.jetbrains.kotlinx:kotlinx-datetime${target.toSuffix()}:${AspVersions.datetime}"
 
 @JvmOverloads
-inline fun bouncycastle(module: String, classifier: String = "jdk18on") =
+ fun bouncycastle(module: String, classifier: String = "jdk18on") =
     "org.bouncycastle:$module-$classifier:${AspVersions.Jvm.bouncycastle}"